[FEAT] #214 - 구매 전 여행일정 컨텐츠 뷰 관련 Hilt 사용한 의존성 주입 적용
