--- app/src/main/java/co/kr/bemyplan/data/repository/purchase/preview/PreviewInfoRepository.kt
@@ -1,7 +1,0 @@
-package co.kr.bemyplan.data.repository.purchase.preview
-
-import co.kr.bemyplan.data.entity.purchase.before.ResponsePreviewInfo
-
-interface PreviewInfoRepository {
-    suspend fun getPreviewInfo(postId: Int): ResponsePreviewInfo
-}
