+++ app/src/main/java/co/kr/bemyplan/data/repository/purchase/preview/PreviewRepository.kt
@@ -1,7 +1,9 @@
 package co.kr.bemyplan.data.repository.purchase.preview
 
+import co.kr.bemyplan.data.entity.purchase.before.ResponsePreviewInfo
 import co.kr.bemyplan.data.entity.purchase.before.ResponsePreviewList
 
-interface PreviewListRepository {
+interface PreviewRepository {
+    suspend fun getPreviewInfo(postId: Int): ResponsePreviewInfo
     suspend fun getPreviewList(postId: Int): ResponsePreviewList
 }
